Struct google_fitness1::AggregateBy

pub struct AggregateBy {
    pub data_source_id: Option<String>,
    pub data_type_name: Option<String>,
}

The specification of which data to aggregate.

This type is not used in any activity, and only used as part of another schema.

Fields

data_source_id

A data source ID to aggregate. Mutually exclusive of dataTypeName. Only data from the specified data source ID will be included in the aggregation. The dataset in the response will have the same data source ID.

data_type_name

The data type to aggregate. All data sources providing this data type will contribute data to the aggregation. The response will contain a single dataset for this data type name. The dataset will have a data source ID of derived:com.google.:com.google.android.gms:aggregated