Struct google_fitness1::Application

pub struct Application {
    pub package_name: Option<String>,
    pub version: Option<String>,
    pub name: Option<String>,
    pub details_url: Option<String>,
}

See: google3/java/com/google/android/apps/heart/platform/api/Application.java

This type is not used in any activity, and only used as part of another schema.

Fields

package_name

Package name for this application. This is used as a unique identifier when created by Android applications, but cannot be specified by REST clients. REST clients will have their developer project number reflected into the Data Source data stream IDs, instead of the packageName.

version

Version of the application. You should update this field whenever the application changes in a way that affects the computation of the data.

name

The name of this application. This is required for REST clients, but we do not enforce uniqueness of this name. It is provided as a matter of convenience for other developers who would like to identify which REST created an Application or Data Source.

details_url

An optional URI that can be used to link back to the application.