Struct google_fitness1::Dataset

pub struct Dataset {
    pub min_start_time_ns: Option<String>,
    pub next_page_token: Option<String>,
    pub max_end_time_ns: Option<String>,
    pub data_source_id: Option<String>,
    pub point: Option<Vec<DataPoint>>,
}

A dataset represents a projection container for data points. They do not carry any info of their own. Datasets represent a set of data points from a particular data source. A data point can be found in more than one dataset.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

data sources datasets patch users (request|response)

data sources datasets get users (response)

Fields
min_start_time_ns: Option<String>

The smallest start time of all data points in this possibly partial representation of the dataset. Time is in nanoseconds from epoch. This should also match the first part of the dataset identifier.

next_page_token: Option<String>

This token will be set when a dataset is received in response to a GET request and the dataset is too large to be included in a single response. Provide this value in a subsequent GET request to return the next page of data points within this dataset.

max_end_time_ns: Option<String>

The largest end time of all data points in this possibly partial representation of the dataset. Time is in nanoseconds from epoch. This should also match the first part of the dataset identifier.

data_source_id: Option<String>

The data stream ID of the data source that created the points in this dataset.

point: Option<Vec<DataPoint>>

A partial list of data points contained in the dataset, ordered by largest endTimeNanos first. This list is considered complete when retrieving a small dataset and partial when patching a dataset or retrieving a dataset that is too large to include in a single response.