Struct google_fitness1::Device

pub struct Device {
    pub model: Option<String>,
    pub version: Option<String>,
    pub type_: Option<String>,
    pub uid: Option<String>,
    pub manufacturer: Option<String>,
}

Representation of an integrated device (such as a phone or a wearable) that can hold sensors. Each sensor is exposed as a data source.

The main purpose of the device information contained in this class is to identify the hardware of a particular data source. This can be useful in different ways, including:
- Distinguishing two similar sensors on different devices (the step counter on two nexus 5 phones, for instance) - Display the source of data to the user (by using the device make / model) - Treat data differently depending on sensor type (accelerometers on a watch may give different patterns than those on a phone) - Build different analysis models for each device/version.

This type is not used in any activity, and only used as part of another schema.

Fields

model

End-user visible model name for the device.

version

Version string for the device hardware/software.

type_

A constant representing the type of the device.

uid

The serial number or other unique ID for the hardware. This field is obfuscated when read by any REST or Android client that did not create the data source. Only the data source creator will see the uid field in clear and normal form.

manufacturer

Manufacturer of the product/hardware.