Struct google-fitness1::UserMethodsBuilder

pub struct UserMethodsBuilder<'a, C, NC, A> where C: 'a, NC: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on user resources. It is not used directly, but through the Fitness hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate "google-fitness1" as fitness1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use fitness1::Fitness;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Fitness::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *MethodBuilders*
// like `data_sources_create(...)`, `data_sources_datasets_delete(...)`, `data_sources_datasets_get(...)`, `data_sources_datasets_patch(...)`, `data_sources_get(...)`, `data_sources_list(...)`, `data_sources_patch(...)`, `data_sources_update(...)`, `sessions_delete(...)`, `sessions_list(...)` and `sessions_update(...)`
// to build up your call.
let rb = hub.users();

Methods

impl<'a, C, NC, A> UserMethodsBuilder<'a, C, NC, A>

fn data_sources_datasets_get(&self, user_id: &str, data_source_id: &str, dataset_id: &str) -> UserDataSourceDatasetGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Returns a dataset containing all data points whose start and end times overlap with the specified range of the dataset minimum start time and maximum end time. Specifically, any data point whose start time is less than or equal to the dataset end time and whose end time is greater than or equal to the dataset start time.

fn data_sources_create(&self, request: &DataSource, user_id: &str) -> UserDataSourceCreateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Creates a new data source that is unique across all data sources belonging to this user. The data stream ID field can be omitted and will be generated by the server with the correct format. The data stream ID is an ordered combination of some fields from the data source. In addition to the data source fields reflected into the data source ID, the developer project number that is authenticated when creating the data source is included. This developer project number is obfuscated when read by any other developer reading public data types.

fn data_sources_datasets_delete(&self, user_id: &str, data_source_id: &str, dataset_id: &str) -> UserDataSourceDatasetDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Performs an inclusive delete of all data points whose start and end times have any overlap with the time range specified by the dataset ID. For most data types, the entire data point will be deleted. For data types where the time span represents a consistent value (such as com.google.activity.segment), and a data point straddles either end point of the dataset, only the overlapping portion of the data point will be deleted.

fn data_sources_datasets_patch(&self, request: &Dataset, user_id: &str, data_source_id: &str, dataset_id: &str) -> UserDataSourceDatasetPatchMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Adds data points to a dataset. The dataset need not be previously created. All points within the given dataset will be returned with subsquent calls to retrieve this dataset. Data points can belong to more than one dataset. This method does not use patch semantics.

fn sessions_delete(&self, user_id: &str, session_id: &str) -> UserSessionDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes a session specified by the given session ID.

fn data_sources_get(&self, user_id: &str, data_source_id: &str) -> UserDataSourceGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Returns a data source identified by a data stream ID.

fn data_sources_update(&self, request: &DataSource, user_id: &str, data_source_id: &str) -> UserDataSourceUpdateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates a given data source. It is an error to modify the data source's data stream ID, data type, type, stream name or device information apart from the device version. Changing these fields would require a new unique data stream ID and separate data source.

Data sources are identified by their data stream ID.

fn sessions_list(&self, user_id: &str) -> UserSessionListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists sessions previously created.

fn sessions_update(&self, request: &Session, user_id: &str, session_id: &str) -> UserSessionUpdateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates or insert a given session.

fn data_sources_patch(&self, request: &DataSource, user_id: &str, data_source_id: &str) -> UserDataSourcePatchMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Updates a given data source. It is an error to modify the data source's data stream ID, data type, type, stream name or device information apart from the device version. Changing these fields would require a new unique data stream ID and separate data source.

Data sources are identified by their data stream ID. This method supports patch semantics.

fn data_sources_list(&self, user_id: &str) -> UserDataSourceListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Lists all data sources that are visible to the developer, using the OAuth scopes provided. The list is not exhaustive: the user may have private data sources that are only visible to other developers or calls using other scopes.