Struct google_fitness1::Session

pub struct Session {
    pub modified_time_millis: Option<String>,
    pub end_time_millis: Option<String>,
    pub description: Option<String>,
    pub activity_type: Option<i32>,
    pub application: Option<Application>,
    pub start_time_millis: Option<String>,
    pub active_time_millis: Option<String>,
    pub id: Option<String>,
    pub name: Option<String>,
}

Sessions contain metadata, such as a user-friendly name and time interval information.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

sessions update users (request|response)

Fields
modified_time_millis: Option<String>

A timestamp that indicates when the session was last modified.

end_time_millis: Option<String>

An end time, in milliseconds since epoch, inclusive.

description: Option<String>

A description for this session.

activity_type: Option<i32>

The type of activity this session represents.

application: Option<Application>

The application that created the session.

start_time_millis: Option<String>

A start time, in milliseconds since epoch, inclusive.

active_time_millis: Option<String>

Session active time. While start_time_millis and end_time_millis define the full session time, the active time can be shorter and specified by active_time_millis. If the inactive time during the session is known, it should also be inserted via a com.google.activity.segment data point with a STILL activity value

id: Option<String>

A client-generated identifier that is unique across all sessions owned by this particular user.

name: Option<String>

A human readable name of the session.